cosas

+++ include/Clases/Compra.h
@@ -19,7 +19,7 @@ class Compra
         Cliente* clienteDeCompra;
     public:
         Compra();
-        Compra(Cliente* _clienteDeCompra, int _montoFinal);
+        Compra(Cliente* _clienteDeCompra, float _montoFinal);
         ~Compra(); //Destructor
 
 
@@ -30,7 +30,7 @@ class Compra
         string getNickCliente();
         /*======= SETTERS ========*/
         void setFecha(DTFecha f);
-        void setMontoFinal(int _montoFinal);
+        void setMontoFinal(float _montoFinal);
         void setCliente(Cliente* _cliente);
 
 };
